+++ locadora.diurno.bll/src/locadora/diurno/bll/CorEJB.java
@@ -1,14 +1,15 @@
 package locadora.diurno.bll;
 
-import locadora.diurno.bll.interfaces.*;	
-import locadora.diurno.bll.util.*;
-import locadora.diurno.dal.dao.interfaces.ICorDAO;
-import locadora.diurno.dal.dao.interfaces.IModeloDAO;
-import locadora.diurno.dal.entidade.*;
-import java.util.*;
+import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+
+import locadora.diurno.bll.interfaces.ICorEJB;
+import locadora.diurno.bll.util.Mensagem;
+import locadora.diurno.bll.util.TipoMensagem;
+import locadora.diurno.dal.dao.interfaces.ICorDAO;
+import locadora.diurno.dal.entidade.Cor;
 
 @Stateless
 public class CorEJB implements ICorEJB{
@@ -18,49 +19,50 @@ public class CorEJB implements ICorEJB{
 	
 	@Override
 	public Mensagem salvar(Cor cor) {
-
-		try {
-			corDAO.insertOrUpdate(cor);
-		}catch(Exception ex) {
-			return new Mensagem("Ocorreu um erro inesperado: " 
-						+ ex.getMessage(),MensagemStatus.erro);
-		}
 		
-		return new Mensagem("Salvo com sucesso.", MensagemStatus.sucesso);
+		try {
+			
+			corDAO.insertOrUpdate(cor);
+			
+			return new Mensagem("Salvo com sucesso.",
+					TipoMensagem.sucesso);
+			
+		}catch(Exception ex) {
+			
+			return new Mensagem("Erro inesperado: " 
+					+ ex.getMessage(), TipoMensagem.erro);
+			
+		}
 	}
 
 	@Override
 	public Mensagem excluir(Short idCor) {
 		
-		
 		try {
 			
-			Cor cor = obterPorId(idCor);
+			Cor cor = corDAO.findById(idCor);
 			
 			if(cor == null) {
-				throw new Exception("Cor inexistente.");
+				throw new Exception("Cor já foi excluída.");
+			}
+			
+			if(cor.getAutomoveis().size() > 0) {
+				throw new Exception("Essa cor possui automoveis vinculados");
 			}
 			
 			corDAO.delete(cor);
 			
+			return new Mensagem("Cor excluída.", TipoMensagem.sucesso);
 		}catch(Exception ex) {
-			return new Mensagem("Não foi possível excluir: " 
-					+ ex.getMessage(), MensagemStatus.erro);
+			return new Mensagem("Erro inesperado: "
+					+ ex.getMessage(), TipoMensagem.erro);
 		}
 		
-		return new Mensagem("Excluído com sucesso.",
-				MensagemStatus.sucesso);
-		
 	}
 
 	@Override
-	public Cor obterPorId(Short idCor) {
-		return corDAO.findById(idCor);
-	}
-
-
-	@Override
-	public List<Cor> obterTodos() {
+	public List<Cor> listar() {
 		return corDAO.findAll();
 	}
+
 }
